+++ src/notifierMixin.ts
@@ -22,12 +22,6 @@ export class AdvancedNotifierNotifierMixin extends SettingsMixinDeviceBase<any> 
             defaultValue: true,
             immediate: true,
         },
-        addSnoozeActions: {
-            title: 'Add snoozing actions',
-            type: 'boolean',
-            defaultValue: true,
-            immediate: true,
-        },
         postNotificationWebhook: {
             subgroup: 'Webhooks',
             type: 'html',